Tighten test-support import locking (#60)
Route the standalone `msgspec` smoke-test import through `with_python`
and restore absent `msgspec` entries by removing the temporary test stub.

Split Python monkeypatch helpers into focused nested modules so each test
support source file stays below the project line-count limit, and expand
module docs for the bootstrap and div BDD test support modules.
